requirements.txt
fastapi
uvicorn
youtube-transcript-api
google-generativeai
duckduckgo-search
gradio
python-dotenv
requests
apify-client
